requirements.txt
ExifRead==2.1.2
gpsphoto==2.2.3
piexif==1.1.3
Pillow==6.2.1
